Split attachment tests by export
Replace the four-thousand-line attachment suite with discoverable per-export files while sharing only the common DOM fixture helper.
