+++ tests/vitest/attachments/float.test.ts
@@ -1,0 +1,87 @@
+import { float } from '$lib/attachments'
+import { afterEach, describe, expect, it, vi } from 'vite-plus/test'
+import { create_element, mock_rect, stub_prop } from '../index'
+
+describe(`float`, () => {
+  const cleanups: (() => void)[] = []
+  afterEach(() => cleanups.splice(0).forEach((cleanup) => cleanup()))
+
+  // anchor as a bare rect, so no element geometry has to be mocked for the anchor
+  const anchor_rect = { top: 100, bottom: 140, left: 60, right: 200 }
+
+  const attach_float = (options: Parameters<typeof float>[0] = {}) => {
+    const node = create_element()
+    mock_rect(node, { left: 0, top: 0, width: 50, height: 20 })
+    const cleanup = float({ anchor: anchor_rect, ...options })(node)
+    if (cleanup) cleanups.push(cleanup)
+    return node
+  }
+
+  it.each([
+    [`fixed`, `fixed`, 0, 0],
+    // absolute is measured against the document, so page scroll has to be added back
+    [`absolute`, `absolute`, 400, 700],
+  ] as const)(
+    `%s strategy positions relative to the right origin`,
+    (_desc, strategy, scroll_x, scroll_y) => {
+      cleanups.push(
+        stub_prop(globalThis, `scrollX`, scroll_x),
+        stub_prop(globalThis, `scrollY`, scroll_y),
+      )
+
+      const node = attach_float({ strategy, placement: `bottom`, align: `start` })
+
+      expect(node.style.position).toBe(strategy)
+      // bottom placement sits below the anchor, start aligns the left edges
+      expect(node.style.top).toBe(`${140 + scroll_y}px`)
+      expect(node.style.left).toBe(`${60 + scroll_x}px`)
+    },
+  )
+
+  it(`match_width sets the exact border-box width and restores inline sizing`, () => {
+    const matched = attach_float({ match_width: true })
+    expect([
+      matched.style.width,
+      matched.style.minWidth,
+      matched.style.boxSizing,
+    ]).toEqual([`140px`, `140px`, `border-box`])
+    expect(attach_float().style.width).toBe(``)
+
+    const node = create_element()
+    node.style.cssText = `box-sizing: content-box; min-width: 10rem; width: 20px`
+    const original_sizing = [node.style.boxSizing, node.style.minWidth, node.style.width]
+    mock_rect(node, { left: 0, top: 0, width: 50, height: 20 })
+    const cleanup = float({ anchor: anchor_rect, match_width: true })(node)
+    cleanup?.()
+    expect([node.style.boxSizing, node.style.minWidth, node.style.width]).toEqual(
+      original_sizing,
+    )
+  })
+
+  it(`repositions on scroll using the floating element's window`, () => {
+    const animation_host = Object.assign(new EventTarget(), {
+      requestAnimationFrame: vi.fn((callback: FrameRequestCallback) => {
+        callback(0)
+        return 42
+      }),
+      cancelAnimationFrame: vi.fn(),
+    }) as unknown as Window
+    const node = create_element()
+    mock_rect(node, { left: 0, top: 0, width: 50, height: 20 })
+    cleanups.push(stub_prop(node, `ownerDocument`, { defaultView: animation_host }))
+    const cleanup = float({ anchor: anchor_rect, placement: `bottom` })(node)
+    node.style.top = `0px`
+    animation_host.dispatchEvent(new Event(`scroll`))
+    expect(node.style.top).toBe(`140px`)
+    cleanup?.()
+  })
+
+  it.each([
+    [`disabled`, { enabled: false }],
+    [`no anchor`, { anchor: null }],
+  ] as const)(`%s attaches nothing`, (_desc, options) => {
+    const node = create_element()
+    expect(float({ anchor: anchor_rect, ...options })(node)).toBeUndefined()
+    expect(node.style.position).toBe(``)
+  })
+})
